+++ src/main/java/com/enviro/assessment/grad001/SinethembaVilakazimain/Service/RecycleTipServiceImpl.java
@@ -7,6 +7,10 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
+
+/**
+ * Implementation of RecyclingTipService interface that provides methods to manage RecyclingTip entities.
+ */
 @Service
 public class RecycleTipServiceImpl implements RecyclingTipService{
     @Autowired
@@ -14,8 +18,10 @@ public class RecycleTipServiceImpl implements RecyclingTipService{
 
 
     /**
-     * @param recyclingTip
-     * @return
+     * Create a new recycling tip.
+     *
+     * @param recyclingTip The recycling tip to create.
+     * @return The created recycling tip.
      */
     @Override
     public RecyclingTip createRecyclingTip(RecyclingTip recyclingTip) {
@@ -23,8 +29,10 @@ public class RecycleTipServiceImpl implements RecyclingTipService{
     }
 
     /**
-     * @param id
-     * @return
+     * Get a recycling tip by ID.
+     *
+     * @param id The ID of the recycling tip to retrieve.
+     * @return An Optional containing the recycling tip if found, otherwise empty.
      */
     @Override
     public Optional<RecyclingTip> getRecyclingTipById(Long id) {
@@ -32,7 +40,9 @@ public class RecycleTipServiceImpl implements RecyclingTipService{
     }
 
     /**
-     * @return
+     * Get all recycling tips.
+     *
+     * @return A list of all recycling tips.
      */
     @Override
     public List<RecyclingTip> getAllRecyclingTips() {
@@ -40,9 +50,11 @@ public class RecycleTipServiceImpl implements RecyclingTipService{
     }
 
     /**
-     * @param id
-     * @param recyclingTip
-     * @return
+     * Update an existing recycling tip.
+     *
+     * @param id            The ID of the recycling tip to update.
+     * @param recyclingTip  The updated recycling tip object.
+     * @return The updated recycling tip.
      */
     @Override
     public RecyclingTip updateRecyclingTip(Long id, RecyclingTip recyclingTip) {
@@ -50,17 +62,31 @@ public class RecycleTipServiceImpl implements RecyclingTipService{
             recyclingTip.setId(id);
             return recycleTipRepository.save(recyclingTip);
         }
-        return null;
+        return null; // Return null if recycling tip with the given ID does not exist
     }
 
     /**
-     * @param id
+     * Delete a recycling tip by ID.
+     *
+     * @param id The ID of the recycling tip to delete.
      */
     @Override
-    public void deleteRecyclingTip(Long id) {
-        recycleTipRepository.deleteById(id);
+    public  boolean deleteRecyclingTip(Long id) {
+        Optional<RecyclingTip> optionalRecyclingTip = recycleTipRepository.findById(id);
+        if (optionalRecyclingTip.isPresent()) {
+            recycleTipRepository.deleteById(id);
+            return true; // Deletion successful
+        } else {
+            return false; // Category with the given ID not found
+        }
     }
 
+    /**
+     * Get all recycling tips for a specific category.
+     *
+     * @param categoryName The category name for which recycling tips are required.
+     * @return A list of recycling tips for the specified category.
+     */
     @Override
     public List<RecyclingTip> getRecyclingTipsByCategoryName(String categoryName) {
         return recycleTipRepository.findByCategoryName(categoryName);
